Remove toolchain include
When you specify toolchain file, cmake makes the variables from it available to all CMakeLists, that is cmake includes this file for you if you specifiy -DCMAKE_TOOLCHAIN_FILE.

This also breaks crosscompilation for me when I use this library, when removed everything works fine.

What is the reasoning behind this? Am I missing something?