:wrench: Fix build with OPENSSL_NO_DEPRECATED (#228)
* Add missing include to fix build with OPENSSL_NO_DEPRECATED

Creating and verifying RSA signatures requires some OpenSSL functions
declared in the `openssl/rsa.h` header. When `OPENSSL_NO_DEPRECATED` is
not defined this header gets indirectly included. But with this define
set the function declarations are missing. This commit adds an explicit
include for the file to fix this.

* Remove unused functions in OpenSSLErrorTest

Some functions from OpenSSL redefined in `OpenSSLErrorTest.cpp` use
types that are not available when `OPENSSL_NO_DEPRECATED` is defined.
Since they do not seem to be actually used this commit simply removes
them.

* Test with OPENSSL_NO_DEPRECATED

Co-authored-by: Chris Mc <prince.chrismc@gmail.com>