Allow to retrieve list of claims
Currently, it is not possible to iterate over the list of claims of a JWK. With this commit, the JWK gains a new method that exposes the underlying claims so that the callee can do some processing with it. This might be useful, if someone wants to copy all claims to a new object.